requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Forms
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Environment Management
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# Development Tools (optional)
Flask-DebugToolbar==0.14.1

# Production Server (optional)
gunicorn==21.2.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0